updated: using thirdweb for wallet management, deploying, refactor client
updated:
- using thirdweb for wallet management, deploying
- refactor client with provider pattern and layout, next 14 error handling
- added: base chain support
- deps update
