+++ client/.env.example
@@ -1,1 +1,2 @@
-NEXT_PUBLIC_COMPILER_ROUTE=http://localhost:4000/api/compile
+NEXT_PUBLIC_COMPILER_ENDPOINT=http://localhost:4000/api/compile
+NEXT_PUBLIC_THIRDWEB_CLIENT_ID=YOUR_THIRDWEB_CLIENT_ID
