+++ client/app/layout.jsx
@@ -1,5 +1,7 @@
 import './globals.css';
 import { Inter } from 'next/font/google';
+import SiteLayout from "./components/SiteLayout";
+import Web3Provider from "./components/Web3Provider";
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -11,7 +13,13 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Web3Provider>
+          <SiteLayout>
+            {children}
+          </SiteLayout>
+        </Web3Provider>
+      </body>
     </html>
   );
 }
