+++ client/app/page.module.css
@@ -1,37 +1,6 @@
-.navbar {
-  position: sticky;
-  top: 0;
-  z-index: 2;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px;
-  background-color: #f0f0f0;
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-}
-
-.logoImage {
-  width: 40px;
-  height: 40px;
-  margin-right: 10px;
-}
-
 .remixLogo {
   width: 15px;
   height: 15px;
-}
-
-.logoText {
-  font-size: 18px;
-}
-
-.navbarButtons {
-  display: flex;
-  gap: 5px;
 }
 
 .container {
